Add file backends and extractors
