+++ solution/text-extractor/backends/exceptions.py
@@ -1,0 +1,6 @@
+class BackendInitException(Exception):
+    pass
+
+
+class BackendException(Exception):
+    pass
