+++ solution/text-extractor/extractors/exceptions.py
@@ -1,2 +1,5 @@
 class ExtractorInitException(Exception):
     pass
+
+class ExtractorException(Exception):
+    pass
