+++ solution/text-extractor/extractors/text.py
@@ -1,4 +1,5 @@
 from . import Extractor
 
+
 class TextExtractor(Extractor):
     file_types = ("text/plain",)
